refactor: changed file position and name
